+++ packages/amplify-appsync-simulator/src/schema/appsync-scalars/index.ts
@@ -1,5 +1,5 @@
 import { URL } from 'url';
-import { GraphQLInt, GraphQLScalarType, GraphQLError, Kind, StringValueNode } from 'graphql';
+import { GraphQLInt, GraphQLScalarType, GraphQLError, Kind, StringValueNode, ValueNode } from 'graphql';
 import { isValidNumber } from 'libphonenumber-js';
 
 import { GraphQLDate, GraphQLTime, GraphQLDateTime } from 'graphql-iso-date';
@@ -104,8 +104,12 @@ are also accepted and these represent the number of seconds till 1970-01-01T00:0
   parseValue(value) {
     return GraphQLInt.parseValue(value) ? value : undefined;
   },
-  parseLiteral(value) {
-    return GraphQLInt.parseLiteral(value, null) ? (value as StringValueNode).value : undefined;
+  parseLiteral(value: ValueNode) {
+    if (value.kind !== Kind.INT) {
+      throw new GraphQLError(`Can only validate integers but received: ${value.kind}`);
+    }
+
+    return Number.parseInt(value.value, 10);
   },
 });