+++ packages/amplify-cli/src/domain/amplify-usageData/IUsageData.ts
@@ -1,9 +1,10 @@
 import { Input } from '../input';
+import { ProjectSettings } from './UsageDataPayload';
 
 export interface IUsageData {
   emitError(error: Error): Promise<void>;
   emitInvoke(): Promise<void>;
   emitAbort(): Promise<void>;
   emitSuccess(): Promise<void>;
-  init(installationUuid: String, version: String, input: Input): void;
+  init(installationUuid: string, version: string, input: Input, accountId: string, projectSettings: ProjectSettings): void;
 }