+++ packages/amplify-cli/src/domain/amplify-usageData/NoUsageData.ts
@@ -13,7 +13,8 @@ export class NoUsageData implements IUsageData {
   emitSuccess(): Promise<void> {
     return Promise.resolve();
   }
-  init(installationUuid: String, version: String, input: any): void {}
+
+  init(installationUuid: string, version: String, input: any, accountId: string): void {}
 
   private static instance: NoUsageData;
   static get Instance(): IUsageData {