+++ packages/amplify-cli/src/domain/amplify-usageData/VersionManager.ts
@@ -1,6 +1,6 @@
 import semver from 'semver';
 
-const APIVersionToPayloadVersion = new Map<string, Array<string>>([['v1.0', ['1.0.0', '1.0.1']]]);
+const APIVersionToPayloadVersion = new Map<string, Array<string>>([['v1.0', ['1.0.0', '1.0.1', '1.1.0']]]);
 
 export function getLatestApiVersion(): string {
   return [...APIVersionToPayloadVersion.keys()].reduce(getMaxVersion, '0');