+++ packages/amplify-e2e-core/src/configure/index.ts
@@ -13,7 +13,7 @@ const defaultSettings = {
   userName: '\r',
 };
 
-const regionOptions = [
+export const amplifyRegions = [
   'us-east-1',
   'us-east-2',
   'us-west-2',
@@ -25,10 +25,12 @@ const regionOptions = [
   'ap-southeast-1',
   'ap-southeast-2',
   'ap-south-1',
+  'ca-central-1',
 ];
 
-const configurationOptions = ['project', 'profile', 'containers'];
-const profileOptions = ['cancel', 'update', 'remove'];
+const configurationOptions = ['Project information', 'AWS Profile setting', 'Advanced: Container-based deployments'];
+const profileOptions = ['No', 'Update AWS Profile', 'Remove AWS Profile'];
+const authenticationOptions = ['AWS profile', 'AWS access keys'];
 
 const MANDATORY_PARAMS = ['accessKeyId', 'secretAccessKey', 'region'];
 
@@ -46,7 +48,7 @@ export function amplifyConfigure(settings: AmplifyConfiguration): Promise<void> 
       .sendCarriageReturn()
       .wait('Specify the AWS Region');
 
-    singleSelect(chain, s.region, regionOptions);
+    singleSelect(chain, s.region, amplifyRegions);
 
     chain
       .wait('user name:')
@@ -72,18 +74,53 @@ export function amplifyConfigure(settings: AmplifyConfiguration): Promise<void> 
   });
 }
 
-export function amplifyConfigureProject(settings: { cwd: string; enableContainers: boolean }): Promise<void> {
-  const { enableContainers = false, cwd } = settings;
+// TODO amplify admin enabled case
+export function amplifyConfigureProject(settings: {
+  cwd: string;
+  enableContainers?: boolean;
+  configLevel?: string;
+  profileOption?: string;
+  authenticationOption?: string;
+  region?: string;
+}): Promise<void> {
+  const {
+    cwd,
+    enableContainers = false,
+    profileOption = profileOptions[0],
+    authenticationOption,
+    configLevel = 'project',
+    region = defaultSettings.region,
+  } = settings;
 
   return new Promise((resolve, reject) => {
     const chain = spawn(getCLIPath(), ['configure', 'project'], { cwd, stripColors: true }).wait('Which setting do you want to configure?');
+
     if (enableContainers) {
-      singleSelect(chain, 'containers', configurationOptions);
-      chain.wait('Do you want to enable container-based deployments?').sendLine('y');
+      singleSelect(chain, configurationOptions[2], configurationOptions);
+      chain.wait('Do you want to enable container-based deployments?').sendConfirmYes();
     } else {
-      singleSelect(chain, 'profile', configurationOptions);
-      chain.wait('Do you want to update or remove the project level AWS profile?');
-      singleSelect(chain, profileOptions[0], profileOptions);
+      singleSelect(chain, configurationOptions[1], configurationOptions);
+
+      if (configLevel === 'project') {
+        chain.wait('Do you want to update or remove the project level AWS profile?');
+        singleSelect(chain, profileOption, profileOptions);
+      } else {
+        chain.wait('Do you want to set the project level configuration').sendConfirmYes();
+      }
+
+      if (profileOption === profileOptions[1] || configLevel === 'general') {
+        chain.wait('Select the authentication method you want to use:');
+        singleSelect(chain, authenticationOption, authenticationOptions);
+
+        if (authenticationOption === authenticationOptions[0]) {
+          chain.wait('Please choose the profile you want to use').sendCarriageReturn(); // Default profile
+        } else if (authenticationOption === authenticationOptions[1]) {
+          chain.wait('accessKeyId:').sendLine('fake-access-key-id');
+          chain.wait('secretAccessKey:').sendLine('fake-secret-access-key');
+          chain.wait('region:');
+          singleSelect(chain, region, amplifyRegions);
+        }
+      }
     }
 
     chain.wait('Successfully made configuration changes to your project.').run((err: Error) => {