+++ packages/amplify-category-analytics/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.21.3](https://github.com/aws-amplify/amplify-cli/compare/amplify-category-analytics@2.21.2...amplify-category-analytics@2.21.3) (2021-04-19)
+
+**Note:** Version bump only for package amplify-category-analytics
+
+
+
+
+
+## [2.21.2](https://github.com/aws-amplify/amplify-cli/compare/amplify-category-analytics@2.21.1...amplify-category-analytics@2.21.2) (2021-04-14)
+
+**Note:** Version bump only for package amplify-category-analytics
+
+
+
+
+
+## [2.21.1](https://github.com/aws-amplify/amplify-cli/compare/amplify-category-analytics@2.20.5...amplify-category-analytics@2.21.1) (2021-04-09)
+
+**Note:** Version bump only for package amplify-category-analytics
+
+
+
+
+
 ## [2.20.5](https://github.com/aws-amplify/amplify-cli/compare/amplify-category-analytics@2.20.4...amplify-category-analytics@2.20.5) (2021-03-11)
 
 **Note:** Version bump only for package amplify-category-analytics