+++ packages/amplify-graphql-types-generator/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.8.1](https://github.com/aws-amplify/amplify-cli/compare/amplify-graphql-types-generator@2.7.0...amplify-graphql-types-generator@2.8.1) (2021-04-09)
+
+
+### Bug Fixes
+
+* **cli:** use more inclusive language ([#6919](https://github.com/aws-amplify/amplify-cli/issues/6919)) ([bb70464](https://github.com/aws-amplify/amplify-cli/commit/bb70464d6c24fa931c0eb80d234a496d936913f5))
+
+
+
+
+
 # [2.7.0](https://github.com/aws-amplify/amplify-cli/compare/amplify-graphql-types-generator@2.6.1...amplify-graphql-types-generator@2.7.0) (2021-02-10)