+++ packages/graphql-connection-transformer/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.21.3](https://github.com/aws-amplify/amplify-cli/compare/graphql-connection-transformer@4.21.2...graphql-connection-transformer@4.21.3) (2021-04-19)
+
+**Note:** Version bump only for package graphql-connection-transformer
+
+
+
+
+
+## [4.21.2](https://github.com/aws-amplify/amplify-cli/compare/graphql-connection-transformer@4.21.1...graphql-connection-transformer@4.21.2) (2021-04-14)
+
+**Note:** Version bump only for package graphql-connection-transformer
+
+
+
+
+
+## [4.21.1](https://github.com/aws-amplify/amplify-cli/compare/graphql-connection-transformer@4.20.7...graphql-connection-transformer@4.21.1) (2021-04-09)
+
+**Note:** Version bump only for package graphql-connection-transformer
+
+
+
+
+
 ## [4.20.7](https://github.com/aws-amplify/amplify-cli/compare/graphql-connection-transformer@4.20.6...graphql-connection-transformer@4.20.7) (2021-03-23)
 
 **Note:** Version bump only for package graphql-connection-transformer